+++ Automation/ScriptingEngine/ScriptableComponents/Configurator.cs
@@ -3,17 +3,22 @@
 // License: Public Domain
 
 using Bindito.Core;
+using IgorZ.TimberDev.Utils;
 
 namespace IgorZ.Automation.ScriptingEngine.ScriptableComponents;
 
 // ReSharper disable once UnusedType.Global
 [Context("Game")]
 sealed class Configurator : IConfigurator {
+  static readonly string PatchId = typeof(Configurator).AssemblyQualifiedName;
+
   public void Configure(IContainerDefinition containerDefinition) {
+    HarmonyPatcher.PatchRepeated(PatchId, typeof(FloodgatePatch));
+
     // The order of bindings is npt important for the scripting engine, but in the constructor the siganls/ and actions
     // will be shown in the order of bindings.
 
-    // The building specific components. Order them from the most to the less frequently needed. 
+    // The building-specific components. Order them from the most to the less frequently needed. 
     containerDefinition.Bind<FloodgateScriptableComponent>().AsSingleton();
     containerDefinition.Bind<PausableScriptableComponent>().AsSingleton();
     containerDefinition.Bind<InventoryScriptableComponent>().AsSingleton();