+++ Automation/CHANGELOG.md
@@ -1,4 +1,4 @@
-# v2.1.0 (pre-release):
+# v2.1.0 (12 May 2025):
 * [Feature] Buildings can now emit their state as global signals to be consumed by other buildings.
 * [Feature] New scriptable components `Prioritizable` and `Constructable`.
 * [Feature] New template tool that allows to prioritize by haulers a newly constructed building.